Moving the fault structs for the fault controller to nvme resources
