+++ vm/devices/storage/disk_nvme/nvme_driver/Cargo.toml
@@ -37,6 +37,7 @@ scsi_buffers.workspace = true
 test_with_tracing.workspace = true
 user_driver_emulated_mock.workspace = true
 nvme_test.workspace = true
+nvme_resources.workspace = true
 async-trait.workspace = true
 
 guid.workspace = true
