+++ vm/devices/storage/disk_nvme/nvme_driver/src/tests.rs
@@ -9,12 +9,12 @@ use guid::Guid;
 use inspect::Inspect;
 use inspect::InspectMut;
 use nvme::NvmeControllerCaps;
+use nvme_resources::fault::FaultConfiguration;
+use nvme_resources::fault::QueueFaultBehavior;
 use nvme_spec::Cap;
 use nvme_spec::Command;
 use nvme_spec::Completion;
 use nvme_spec::nvm::DsmRange;
-use nvme_test::FaultConfiguration;
-use nvme_test::QueueFaultBehavior;
 use pal_async::DefaultDriver;
 use pal_async::async_test;
 use pal_async::timer::PolledTimer;
