+++ vm/devices/storage/nvme_resources/Cargo.toml
@@ -8,6 +8,8 @@ rust-version.workspace = true
 
 [dependencies]
 vm_resource.workspace = true
+nvme_spec.workspace = true
+async-trait.workspace = true
 
 mesh.workspace = true
