+++ vm/devices/storage/nvme_test/src/workers/admin.rs
@@ -8,12 +8,10 @@ use super::MAX_DATA_TRANSFER_SIZE;
 use super::io::IoHandler;
 use super::io::IoState;
 use crate::DOORBELL_STRIDE_BITS;
-use crate::FaultConfiguration;
 use crate::MAX_QES;
 use crate::NVME_VERSION;
 use crate::PAGE_MASK;
 use crate::PAGE_SIZE;
-use crate::QueueFaultBehavior;
 use crate::VENDOR_ID;
 use crate::error::CommandResult;
 use crate::error::NvmeError;
@@ -33,6 +31,8 @@ use futures_concurrency::future::Race;
 use guestmem::GuestMemory;
 use guid::Guid;
 use inspect::Inspect;
+use nvme_resources::fault::FaultConfiguration;
+use nvme_resources::fault::QueueFaultBehavior;
 use pal_async::task::Spawn;
 use pal_async::task::Task;
 use parking_lot::Mutex;
