+++ vm/devices/storage/nvme_test/src/workers/coordinator.rs
@@ -8,7 +8,6 @@ use super::admin::AdminConfig;
 use super::admin::AdminHandler;
 use super::admin::AdminState;
 use super::admin::NsidConflict;
-use crate::FaultConfiguration;
 use crate::queue::DoorbellRegister;
 use disk_backend::Disk;
 use futures::FutureExt;
@@ -21,6 +20,7 @@ use inspect::InspectMut;
 use mesh::rpc::PendingRpc;
 use mesh::rpc::Rpc;
 use mesh::rpc::RpcSend;
+use nvme_resources::fault::FaultConfiguration;
 use pal_async::task::Spawn;
 use pal_async::task::Task;
 use parking_lot::Mutex;
